Services and stuff

+++ src/main/java/no/hvl/FeedApp/repositories/PollRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PollRepo extends JpaRepository<Poll, Integer> {
+public interface PollRepo extends JpaRepository<Poll, Long> {
 }
